requirements.txt
Django>=5.0,<6.0
wagtail>=6.2,<7.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-taggit>=5.0,<6.0
django-treebeard>=4.7,<5.0
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
